+++ src/components/dropdown/dropdown.js
@@ -152,6 +152,9 @@ class Dropdown extends React.Component {
     this.setState({ open: false });
     this._handleContentChange();
     this.emitOnChangeCallback(val, visibleVal);
+    if (this.props.onBlur) {
+      this.props.onBlur();
+    }
   }
 
   /**
@@ -233,6 +236,10 @@ class Dropdown extends React.Component {
   handleBlur = () => {
     if (!this.blockBlur) {
       this.setState({ open: false });
+
+      if (this.props.onBlur) {
+        this.props.onBlur();
+      }
     }
   }