+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.34.2
+
+## Bug fix
+
+* Fixes onBlur prop passed to `Date`, `Decimal`, `Dropdown`, `DropdownFilter`, and `DropdownFilterAjax` components so it is called instead of ignored
+
 # 0.34.1
 
 ## Component Enhancements